requirements.txt
requests==2.25.1
jpype1
boto3
pyyaml


